Add example to .magento.env.yaml, move upload to PostDeploy
